+++ pybo/models.py
@@ -2,12 +2,14 @@ from django.db import models
 from django.contrib.auth.models import User
 
 class Question(models.Model):
+    author = models.ForeignKey(User, on_delete=models.CASCADE, related_name='author_question')
+
     subject = models.CharField('제목', max_length=200, help_text="질문의 제목은 한줄로 작성해라")
     content = models.TextField('내용', help_text="침착하게 내용을 작성하세요")
     create_date = models.DateTimeField('작성일')
-    author = models.ForeignKey(User, on_delete=models.CASCADE, null=True)
-
     modify_date = models.DateTimeField(null=True, blank=True)
+
+    voter = models.ManyToManyField(User, related_name='voter_question') # author의 User 모델과 중복되기 때문에 오류가난다. 별명인 voter_question을 주어 이름 중복 문제를 해결
 
     def __str__(self):
         return self.subject
