+++ pybo/urls.py
@@ -14,26 +14,28 @@ Including another URLconf
     2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
 """
 from django.urls import path
-from . import views
+from .views import base_views, question_views, answer_views, comment_views
 
 app_name = "pybo"
 
 urlpatterns = [
-path('', views.index, name='index'),
-path('<int:question_id>/', views.detail, name='detail'),
+path('', base_views.index, name='index'),
+path('<int:question_id>/', base_views.detail, name='detail'),
 
-path('question/create/', views.question_create, name='question_create'),
-path('question/modify/<int:question_id>/', views.question_modify, name='question_modify'),
-path('question/delete/<int:question_id>/', views.question_delete, name='question_delete'),
+path('question/create/', question_views.question_create, name='question_create'),
+path('question/modify/<int:question_id>/', question_views.question_modify, name='question_modify'),
+path('question/delete/<int:question_id>/', question_views.question_delete, name='question_delete'),
 
-path('answer/create/<int:question_id>/', views.answer_create, name='answer_create'),
-path('answer/modify/<int:answer_id>/', views.answer_modify, name='answer_modify'),
-path('answer/delete/<int:answer_id>/', views.answer_delete, name='answer_delete'),
+path('answer/create/<int:question_id>/', answer_views.answer_create, name='answer_create'),
+path('answer/modify/<int:answer_id>/', answer_views.answer_modify, name='answer_modify'),
+path('answer/delete/<int:answer_id>/', answer_views.answer_delete, name='answer_delete'),
 
-path('comment/create/question/<int:question_id>/', views.comment_create_question, name='comment_create_question'),
-path('comment/modify/question/<int:comment_id>/', views.comment_modify_question, name='comment_modify_question'),
-path('comment/delete/question/<int:comment_id>/', views.comment_delete_question, name='comment_delete_question'),
-
+path('comment/create/question/<int:question_id>/', comment_views.comment_create_question, name='comment_create_question'),
+path('comment/modify/question/<int:comment_id>/', comment_views.comment_modify_question, name='comment_modify_question'),
+path('comment/delete/question/<int:comment_id>/', comment_views.comment_delete_question, name='comment_delete_question'),
+path('comment/create/answer/<int:answer_id>/', comment_views.comment_create_answer, name='comment_create_answer'),
+path('comment/modify/answer/<int:comment_id>/', comment_views.comment_modify_answer, name='comment_modify_answer'),
+path('comment/delete/answer/<int:comment_id>/', comment_views.comment_delete_answer, name='comment_delete_answer'),
 ]
