--- pybo/views.py
@@ -1,174 +1,0 @@
-# Response의 Body에 들어갈 html 문서 return (렌더링을 통해) : 
-# 이를 위해 템플릿 파일, 템플릿 파일에서 사용할 변수 목록 필요 
-# -> render() 호출 -> render는 Response Body html 리턴
-
-from django.shortcuts import render, get_object_or_404, redirect
-from .models import Question, Answer
-from django.utils import timezone
-from .forms import QuestionForm, AnswerForm, CommentForm
-from django.core.paginator import Paginator
-from django.contrib.auth.decorators import login_required
-from django.contrib import messages
-
-# Create your views here.
-
-def index(request):
-    """pybo 목록 출력"""
-    page = request.GET.get('page', '1')
-
-    question_list = Question.objects.order_by('-create_date')
-
-    paginator = Paginator(question_list, 10)
-    page_obj = paginator.get_page(page)
-    context = {"question_list": page_obj}
-    
-    return render(request, 'pybo/question_list.html', context)
-
-def detail(request, question_id):
-    question =  get_object_or_404(Question, pk=question_id)
-
-    context = {'question' : question}
-    return render(request, 'pybo/question_detail.html', context)
-
-@login_required(login_url='common:login')
-def answer_create(request, question_id):
-    """pybo 답변등록"""
-
-    question = get_object_or_404(Question, pk=question_id)
-    if request.method == 'POST': # post 방식일 때
-        form = AnswerForm(request.POST)
-        if form.is_valid():
-            answer = form.save(commit=False)
-            answer.author = request.user # 로그인 사용자 정보로 설정
-            answer.create_date = timezone.now()
-            answer.question = question # Foreign Key
-            answer.save()
-            return redirect('pybo:detail', question_id=question.id)
-    else: # get 방식일 때
-        form = QuestionForm()
-    context = {'question':question, 'form':form}
-    return render(request, 'pybo/question_detail.html', context)
-
-@login_required(login_url='common:login')
-def question_create(request):
-    """pybo 질문등록"""
-    if request.method == 'POST':
-        form = QuestionForm(request.POST)
-        if form.is_valid():
-            question = form.save(commit=False)
-            question.author = request.user # 로그인 사용자 정보로 설정
-            question.create_date = timezone.now()
-            question.save()
-            return redirect('pybo:index')
-    else:
-        form = QuestionForm()
-        
-    context = {'form': form}
-    return render(request, 'pybo/question_form.html', context)
-
-@login_required(login_url='common:login')
-def question_modify(request, question_id):
-    """질문 수정"""
-    question = get_object_or_404(Question, pk=question_id)
-    if request.user != question.author:
-        messages.error(request, "수정 권한이 없습니다.")
-        return redirect('pybo:detail', question_id=question.id)
-    
-    if request.method == "POST":
-        # 질문 수정을 위한 값 덮어쓰기
-        form = QuestionForm(request.POST, instance=question)
-        if form.is_valid():
-            question = form.save(commit=False)
-            question.author = request.user
-            question.modify_date = timezone.now() # 수정 일시
-            question.save()
-            return redirect('pybo:detail', question_id=question.id)
-    else:
-        #질문 수정 화면에 기존 제목, 내용 반영
-        form = QuestionForm(instance=question)
-        
-    context = {"form": form}
-
-    return render(request, 'pybo/question_form.html', context)
-
-
-@login_required(login_url='common:login')
-def question_delete(request, question_id):
-    """
-    pybo 질문삭제
-    """
-    question = get_object_or_404(Question, pk=question_id)
-    if request.user != question.author:
-        messages.error(request, '삭제권한이 없습니다')
-        return redirect('pybo:detail', question_id=question.id)
-    question.delete()
-    return redirect('pybo:index')
-
-
-@login_required(login_url='common:login')
-def answer_modify(request, answer_id):
-    """
-    pybo 답변수정
-    """
-    answer = get_object_or_404(Answer, pk=answer_id)
-    if request.user != answer.author:
-        messages.error(request, "수정권한이 없습니다")
-        return redirect('pybo:detail', question_id=answer.question.id)
-    
-    if request.method == "POST":
-        form = AnswerForm(request.POST, instance=answer)
-        if form.is_valid():
-            answer = form.save(commit=False)
-            answer.author = request.user
-            answer.modify_date = timezone.now()
-            answer.save()
-            return redirect('pybo:detail', question_id=answer.question.id)
-    else:
-        form = AnswerForm(instance=answer)
-            
-    context = {'answer' : answer, 'form' : form}
-
-    return render(request, 'pybo/answer_form.html', context)
-
-
-@login_required(login_url='common:login')
-def answer_delete(request, answer_id):
-    """
-    pybo 답변삭제
-    """
-    answer = get_object_or_404(Answer, pk=answer_id)
-    if request.user != answer.author:
-        messages.error(request, "삭제권한이 없습니다")
-    else:
-        answer.delete()
-    return redirect('pybo:detail', question_id=answer.question.id)
-    
-@login_required(login_url='common:login')
-def comment_create_question(request, question_id):
-    question = get_object_or_404(Question, pk=question_id)
-    
-    if request.method == "POST":
-        form = CommentForm(request.POST)
-        if form.is_valid():
-            comment = form.save(commit=False)
-            comment.author = request.user
-            comment.create_date = timezone.now()
-            comment.question = question
-            comment.save()
-            return redirect('pybo:detail', question_id=question.id)
-    else:
-        form = CommentForm()
-            
-    context = {'form' : form}
-
-    return render(request, 'pybo/comment_form.html', context)
-
-
-@login_required(login_url='common:login')
-def comment_modify_question(request, question_id):
-    pass
-
-
-@login_required(login_url='common:login')
-def comment_delete_question(request, question_id):
-    pass
