+++ pybo/views/__init__.py
@@ -1,0 +1,3 @@
+# Response의 Body에 들어갈 html 문서 return (렌더링을 통해) : 
+# 이를 위해 템플릿 파일, 템플릿 파일에서 사용할 변수 목록 필요 
+# -> render() 호출 -> render는 Response Body html 리턴
