+++ static/style.css
@@ -16,7 +16,7 @@
 }
 
 .comment {
-    border-top: dotted 1px blue;
+    border-top: dotted 1px gray;
     font-size: 0.7em;
 }
