+++ templates/pybo/comment_answer.html
@@ -1,0 +1,25 @@
+    <div class="mx-3">
+      {% if answer.comment_set.count > 0 %}
+      <div class="mt-3">
+      {% for comment in answer.comment_set.all %}
+        <div class="comment py-1 text-muted">
+          <span style="white-space: pre-line;"><b>{{ comment.content }}</b></span>
+          <span>
+            -{{ comment.author }} [{{ comment.create_date | date:'y-m-d H:i' }}]
+            {% if comment.modify_date %}
+            (수정:{{ comment.modify_date }})
+            {% endif %}
+          </span>
+          {% if request.user == comment.author %}
+          <a href="{% url 'pybo:comment_modify_answer' comment.id %}" class="small">수정</a>,
+          <a href="#" class="small delete" data-uri="{% url 'pybo:comment_delete_answer' comment.id %}">삭제</a>
+          {% endif %}
+        </div>
+      {% endfor %}
+      </div>
+      {% endif %}
+      <div class="mb-2">
+          <a href="{% url 'pybo:comment_create_answer' answer.id %}" class="small"><small>댓글 추가 ..</small></a>
+
+      </div>
+    </div>
