+++ templates/pybo/comment_question.html
@@ -5,7 +5,8 @@
       <div class="mt-3">
       {% for comment in question.comment_set.all %}
         <div class="comment py-1 text-muted">
-          <span style="white-space: pre-line;">{{ comment.content }}</span>
+          {% include 'pybo/comment_answer.html' %}
+          <span style="white-space: pre-line;"><b>{{ comment.content }}</b></span>
           <span>
             -{{ comment.author }} [{{ comment.create_date | date:'y-m-d H:i' }}]
             {% if comment.modify_date %}
@@ -14,13 +15,13 @@
           </span>
           {% if request.user == comment.author %}
           <a href="{% url 'pybo:comment_modify_question' comment.id %}" class="small">수정</a>,
-          <a href="{% url 'pybo:comment_delete_question' comment.id %}" class="small delete" data-uri="{% url 'pybo:comment_delete_question' comment.id %}">삭제</a>
+          <a href="#" class="small delete" data-uri="{% url 'pybo:comment_delete_question' comment.id %}">삭제</a>
           {% endif %}
         </div>
       {% endfor %}
       </div>
       {% endif %}
-      <div>
+      <div class="mb-2">
           <a href="{% url 'pybo:comment_create_question' question.id %}" class="small"><small>댓글 추가 ..</small></a>
 
       </div>
