+++ templates/pybo/question_detail.html
@@ -30,7 +30,7 @@
     {% endif %}
 
     <div class="card my-3">
-        <div class="card-body" style="white-space: pre-line;">{{question.content}}</div>
+        <div class="card-body">{{question.content}}</div>
 
         <div class="d-flex justify-content-end">
 
@@ -49,7 +49,9 @@
             <div>{{question.create_date}}</div>
           </div>
         </div>
-        {% include 'pybo/question_comment.html' %}
+
+        {% include 'pybo/comment_question.html' %}
+
         {% if request.user == question.author %}
         <div class="card-footer">
             <a href="{% url "pybo:question_modify" question.id %}" class="btn btn-sm btn-outline-secondary text-info">수정</a>
